back/interpreter/pyproject.toml
[tool.poetry]
name = "interpreter"
version = "0.1.0"
description = "Summarization service for flat-rating"
authors = ["nicolas-ricc <nicolas.riccomini@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
transformers = "^4.44.0"
torch = "^2.4.0"
accelerate = "^0.33.0"
requests = "^2.32.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
